% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_plot_functions.R
\name{actogramPlot}
\alias{actogramPlot}
\title{Actogram for nestbox activity}
\usage{
actogramPlot(x)
}
\arguments{
\item{x}{a data.table as supplied by queryEvents of queryResidents}
}
\value{
a ggplot2 object
}
\examples{
con = dbcon(user = "YOUR_USER_NAME", host = getOption("host") )
x = queryEvents(con = con, box = 256, from = "2014-03-05", to = "2014-06-10", includeFIELD = TRUE)
actogramPlot(x)       #plots immediately
p = actogramPlot(x)   #plot saved as ggplot2 object first
p                     #and plotted in a separate step
ggsave("test_actogramPlot.pdf")
getwd()

#plot breeding pair only
x = queryResidents(con = con, x = x)
x = subset(x, resident == 1)
actogramPlot(x)      

closeCon(con)

}
\author{
LS
}
