% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ELEMENTS_events.R
\name{assign_direction}
\alias{assign_direction}
\title{Assign a direction based on individual ins and outs}
\usage{
assign_direction(x)
}
\arguments{
\item{x}{a data.frame as provided by the function fetch_ins_outs}
}
\value{
a data.frame with directional information
}
\description{
This function assigns the direction based on a decision rule (see details).
}
\details{
To assign directions, the first (non-'00'), and the last (non-'00') light barriers are used.
Additionally, the information whether both light barriers were triggered at any point of the event (= individual passes
through the entrance) is used.

The first light barrier is scored as "1" (outer light barrier), "2" (inner light barrier), or "0" (no directional information; either "00" or "12")

Passes through the entrace are scored as either "0" (no pass) or "2" (pass)

The last light barrier is scored as "1" (outer light barrier), "2" (inner light barrier), or "0" (no directional information; either "00" or "12")

Together this can be interpreted as a three-digit number; for example 122 would that first the outer light barrier was triggered,
then both, and then the inner light barrier only. This is clearly a nestbox entry.

Available directions are IN (nestbox entry), OUT (nestbox exit), BACK (hanging inside at the entrance of the nestbox),
FRONT (hanging outside at the entrance of the nestbox), PASS (pass through the nestbox opening without information about the direction),
NOPASS (no light barriers triggered, or light barriers contradictory; no pass through the nestbox opening).

The current direction definitions are given below:
\itemize{
 \item IN: 122|120|022
 \item OUT: 021|221|220
 \item BACK: 222|202
 \item FRONT: 121|101
 \item PASS: 020
 \item NOPASS: 000|102|201
}
}
\note{
Rule definitions are marked in the code with a "###RULE###".
}
\examples{
see function events
}
\seealso{
\code{\link{events}}, \code{\link{fetch_data}}, \code{\link{remove_faulty_data}},
\code{\link{trim_data}}, \code{\link{fetch_ins_outs}}, \code{\link{concat_events}},
}
\author{
LS
}
