% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_EVENTS_elements_v2.r
\name{assign_direction_v2}
\alias{assign_direction_v2}
\title{Assign direction to different pieces of activity}
\usage{
assign_direction_v2(x, hardware_threshold = 0.05)
}
\arguments{
\item{x}{A data.table as supplied by \link{fetch_ins_outs_v2}.}

\item{hardware_threshold}{= 0.05 Time in seconds within which a changed of outer and inner light barrier is assumed to be "at the same time". This is defined by the hardware settings and should therefore usually not be changed.}
}
\description{
Assign direction to different pieces of activity, e.g. "IN" and "OUT"
}
\details{
See \link{events_v2} for further information. Note that after some minor alterations, the output of this function is returned, if in function \link{events_v2} groups_ins_and_outs = FALSE.
}
\examples{
#Not run
#see help of function events_v2
?events_v2()
}
