% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{basename2box}
\alias{basename2box}
\title{basename to box}
\usage{
basename2box(ff)
}
\description{
extract box from a file path (basename) string
}
\examples{
 
basename2box("/ds/raw_data_kemp/FIELD/Westerholz/SNB/RAWDATA/2016/2016.01.26/011/BOX001.TXT")
basename2box("/ds/raw_data_kemp/FIELD/Westerholz/SNB_v2/RAWDATA/2016/2016.01.26/011/BOX0080.TXT")
}
\author{
MV
}
