% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_box_activity.R
\name{box_activity}
\alias{box_activity}
\title{Box activity}
\usage{
box_activity(lastday, days_before = 1)
}
\arguments{
\item{lastday}{a Date object}
}
\value{
A data.table
}
\description{
Box activity estimate based on light-barrier to transponder ratio
}
\examples{
    
\dontrun{
   x = box_activity()
 }            
}
\author{
MV
}
