% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boxes}
\alias{boxes}
\title{3 digits box list}
\usage{
boxes(x = 1:277)
}
\arguments{
\item{x}{box names (numeric)}
}
\value{
\code{data.table} with column box as character eg. 'b001'
}
\description{
A list of boxes in westerholz formated as b00x
}
\examples{
  boxes()
}
\author{
MV
}
