% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_eventsDataQueries.R
\name{breedingStage}
\alias{breedingStage}
\title{Function that adds hatching and laying start information to a dataset, if present.}
\usage{
breedingStage(con, x)
}
\arguments{
\item{x}{A data.table as returned by the function "feeding". Otherwise, any data.table
with an (integer) year_, an (integer) box, and an IDate date column.}

\item{user}{Your username (as necessary for the connection to the database)}
}
\value{
A data.table where with four additional columns: hatchDate (date of hatching),
firstEgg (date of first Egg), dayToFirstEgg (number of days after start of egg laying),
dayToHatch (number of days after hatching start); the latter two columns are positive for
later dates, and negative for earlier dates.
}
\examples{
con = dbcon(user = "YOUR_USER_NAME", host = getOption("host") )
x = queryEvents(con = con, box = 26, from = "2014-05-17", to = "2014-05-18", includeFIELD = TRUE)
x = queryResidents(con = con, x = x)
x = feeding(x = x, timeScale = 10, from = '2014-05-17', to = '2014-05-18')
x = breedingStage(con = con, x = x)
closeCon(con)
}
\seealso{
\code{\link{feeding}}
}
\author{
LS
}
