% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_butler_INTERFACE.R
\name{butler}
\alias{butler}
\title{SNB butler}
\usage{
butler(user = "testuser", in.browser = FALSE,
  HOST = "scidb.mpio.orn.mpg.de")
}
\arguments{
\item{user}{local system and db user (passed to install_demo_SNB) .}

\item{in.browser}{default to FALSE. If true open the local interface in the default browser.}

\item{demo}{default to FALSE; if TRUE runs  `install_demo_SNB()` first.}
}
\value{
nothing
}
\description{
SNB User Interfaces
}
\examples{
 
\dontrun{       
butler(demo = TRUE)
}
}
\author{
MV
}
