% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_VideoVsEvents.r
\name{calcTimeInside_v2}
\alias{calcTimeInside_v2}
\title{Helper function for function \link{checkIn_v2}}
\usage{
calcTimeInside_v2(x, timescale)
}
\description{
Helper function for function \link{checkIn_v2}
}
