% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_butler_ELEMENTS.R
\name{card_format_status}
\alias{card_format_status}
\title{Checks inserted SD cards.}
\usage{
card_format_status(x = cardReader())
}
\arguments{
\item{x}{a DT as returned by cardReader()}

\item{boxid}{data.table with current box -hwid}
}
\value{
data.table
}
\description{
Checks inserted SD cards.
}
\examples{
card_format_status()
}
\author{
MV
}
