% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_butler_ELEMENTS.R
\name{card_prepare_status}
\alias{card_prepare_status}
\title{Checks inserted SD cards.}
\usage{
card_prepare_status(x = cardReader())
}
\arguments{
\item{x}{a DT as returned by cardReader()}

\item{boxid}{data.table with current box -hwid}
}
\value{
data.table
}
\examples{
card_prepare_status()
}
\author{
MV
}
