% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char2date}
\alias{char2date}
\title{Extract date}
\usage{
char2date(x)
}
\arguments{
\item{x}{character vector(s)}
}
\value{
\code{Date} object
}
\description{
Extract date (yyy.mm.dd) from a character string (e.g. a file path)
}
\examples{
char2date(x = '2015/2015.01.26/001/box001.txt')
}
\author{
MV
}
