% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_VideoVsEvents.r
\name{checkIn_v2}
\alias{checkIn_v2}
\title{Assemble the data for comparing SNB and video data by unit of time.}
\usage{
checkIn_v2(vTCout, timescale = 1)
}
\description{
Assemble the data for comparing SNB and video data by unit of time.
}
\details{
Called by \link{SNBvsVid_v2}, please follow the link for further details.
}
