% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_EVENTS.R
\name{combine_front_v2}
\alias{combine_front_v2}
\title{Groups FRONTs that are close in time.}
\usage{
combine_front_v2(x, cluster_fronts_threshold = 5, no_front = NULL)
}
\arguments{
\item{x}{A data.table as supplied by \link{concat_direction_v2}.}

\item{cluster_fronts_threshold}{Time in seconds within which FRONTs are grouped together.}

\item{no_front}{Define which activities are assumed to NOT be a front. Defaults to anything that is completely unclear, or that ended or started inside of the box (e.g. OUT, IN, BACK, ...)}
}
\description{
Example: A bird pokes its head into the nest opening and pulls it back, this can be scored as different events or as belonging to the same one.
}
\details{
See \link{events_v2} for further information.
}
\examples{
#Not run
#see help of function events_v2
?events_v2()
}
