% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_EVENTS_elements_v1.R
\name{concat_events_v1}
\alias{concat_events_v1}
\title{Combind ins and outs to events}
\usage{
concat_events_v1(x, max_distance)
}
\arguments{
\item{x}{a data.frame as provided by the function assign_direction}
}
\value{
a data.frame ins and outs grouped together to single events, where possible.
}
\description{
This function combines individual instances at the nestbox to nestbox presence and absence data, where possible.
The function aims to be conservative; it aims to combine two (or more) events only if they are extremly likely to belong together.
Functions that can be used for more precise, but also less conservative data can be found elsewhere (tba).
}
\details{
In order to avoid loops, this function defines for each direction available whether before or after this
line a new event should begin. For example, "IN" and "FRONT" always start an event (break added before this line),
OUT always stops an event (break added after this line). "PASSES" may be grouped together with INs and OUTs, but
not with other PASSES. Events where the bird did not pass through the entrance are not considered to begin or end
events per se.
}
\note{
Rule definitions are marked in the code with a "###RULE###".
}
\examples{
see function events
}
\seealso{
\code{\link{events}}, \code{\link{fetch_data}}, \code{\link{remove_faulty_data}},
\code{\link{trim_data}}, \code{\link{fetch_ins_outs}}, \code{\link{assign_direction}},
}
\author{
LS
}
