% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_EVENTS_elements_v2.r
\name{concat_events_v2}
\alias{concat_events_v2}
\title{Groups activity that belongs together.}
\usage{
concat_events_v2(x, max_distance = 14 * 60 * 60)
}
\arguments{
\item{x}{A data.table as supplied by \link{assign_direction_v2}.}

\item{max_distance}{Maximum time in seconds within which different pieces of activity will be considered as belonging to a single event.}
}
\description{
Groups activity that belongs together, e.g. IN-OUT, IN-BACK-OUT, ...
}
\details{
See \link{events_v2} for further information.
}
\examples{
#Not run
#see help of function events_v2
?events_v2()
}
