% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileop.R
\name{data_files}
\alias{data_files}
\title{List all data files}
\usage{
data_files(p = getOption("path.to.raw_v2"))
}
\arguments{
\item{p}{path to raw data, default to getOption("path.to.raw")}
}
\value{
a data.table if there are new directories or otherwise NULL
}
\description{
List all data files that are box.{1,4}.txt$
}
\examples{
\dontrun{
require(SNB2)
}
}
\author{
MV
}
