% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_queries.R
\name{dbqSNB}
\alias{dbqSNB}
\title{query raw data, all b_ tables at once}
\usage{
dbqSNB(username, host = "scidb.mpio.orn.mpg.de",
  q = "SELECT * FROM boxtables limit 1", db = getOption("snbDB_v2"),
  .boxes = getOption("boxes_v2"), ncores = 4)
}
\arguments{
\item{username}{username}

\item{host}{host}

\item{q}{query,boxtables kw should be used instead of bnnn box name.}

\item{db}{default to getOption('snbDB_v2')}

\item{.boxes}{default to 1:277}

\item{ncores}{number of cores to use}
}
\description{
query raw data, all b_ tables at once
}
\examples{
\dontrun{
# Simple SELECT
dbqSNB('snbAdmin', host = "scidb.mpio.orn.mpg.de", 'SELECT * FROM boxtables limit 1')

# Last entry
dbqSNB('snbAdmin', host = "scidb.mpio.orn.mpg.de", 'SELECT * FROM boxtables ORDER BY r_pk desc limit 1')

# SELECT last n days
x = dbqSNB('snbAdmin', host = "scidb.mpio.orn.mpg.de",
'SELECT datetime_, LB, transp FROM  boxtables
   WHERE datetime_ >= DATE_ADD( (SELECT max(datetime_) from  boxtables) , INTERVAL -3 DAY) AND
   YEAR(datetime_) = YEAR(CURDATE() )
')
}

}
