% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fileop.R
\name{demo_setup}
\alias{demo_setup}
\title{Installs a demo system}
\usage{
demo_setup(user = "testuser", host = "127.0.0.1",
  rawdata_root = paste(tempdir(), "SNB_demo_RAWDATA", sep = "/"),
  db = "tests")
}
\description{
Installs a demo system prepares a temp RAWDATA folder and changes the options
}
\examples{
\dontrun{
demo_setup()
}

}
