% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_rawdata_checks.R
\name{diagnose_pull}
\alias{diagnose_pull}
\title{Diagnose directory of one pull including drift if available}
\usage{
diagnose_pull(date, shiny = FALSE)
}
\arguments{
\item{date}{a character vector formated as 'yyyy.mm.dd' (same as the directory date holding the raw data)}

\item{outDirLocation}{the location of the raw data; the default is getOption('path.to.raw')}
}
\value{
a data.table
}
\description{
Output diagnostics of a directory containing raw box files (including )
}
\examples{
x = diagnose_pull(date= "2017.03.03")
}
\author{
MV
}
