% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_rawdata_checks.R
\name{diagnose_pull_v2}
\alias{diagnose_pull_v2}
\title{Diagnose directory of one pull (i.e. a directory)}
\usage{
diagnose_pull_v2(
  date,
  outDirLocation = getOption("path.to.raw_v2"),
  shiny = FALSE
)
}
\arguments{
\item{date}{a character vector formated as 'yyyy.mm.dd' (same as the directory date holding the raw data)}

\item{outDirLocation}{the location of the raw data; the default is getOption('path.to.raw_v2')}
}
\value{
a data.table
}
\description{
Output diagnostics of a directory containing raw box files
}
\examples{
\dontrun{
options(host = 'scidb.mpio.orn.mpg.de')  
x = diagnose_pull_v2(date = "2018.04.03")
}
}
\author{
MV
}
