% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_rawdata_checks.R
\name{diagnose_raw_dir}
\alias{diagnose_raw_dir}
\title{Output diagnostics on row SNB files}
\usage{
diagnose_raw_dir(date, outDirLocation = getOption("path.to.raw"))
}
\arguments{
\item{date}{a character vector formated as 'yyyy.mm.dd' (same as the directory date holding the raw data)}

\item{outDirLocation}{the location of the raw data; the default is getOption('path.to.raw')}
}
\description{
Output diagnostics of a directory containing raw box files
}
\examples{
date = "2016.05.10"
x = diagnose_raw_dir(date)
openxlsx::write.xlsx(x, file = '~/temp.xlsx')
}
\seealso{
\code{\link{diagnose_raw_txt}}
}
\author{
MV
}
