% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_rawdata_checks.R
\name{diagnose_raw_txt}
\alias{diagnose_raw_txt}
\title{Output diagnostics on raw SNB files}
\usage{
diagnose_raw_txt(filePath)
}
\value{
A data.table with the following fields:

\itemize{
 \item{\strong{box                 }} When no information is given for one box then something is definitely wrong with that box
 \item{\strong{file_size           }}
 \item{\strong{n_lines             }} Number of lines in the original text file
 \item{\strong{n_unique_lines      }} Number of unique lines in the file
 \item{\strong{prop_wrong_years    }} Proportion of wrong years
 \item{\strong{first_good_date     }} First correct date
 \item{\strong{last_good_date      }} Last correct date
 \item{\strong{N_transponders      }} Number of detected transponders
 \item{\strong{pos_last_seen_transp}} Position of the last observed (test) transponder (0 means seen on the last line)
 \item{\strong{last_batvolt        }} Last recorded battery voltage. No value means no voltage data.
 \item{\strong{min_batvolt         }} Lower recorded battery voltage. No value means no voltage data.
 \item{\strong{bat_discharge_rate  }} Battery discharge daily rate. No value means no voltage data.
}
}
\description{
Output diagnostics of a raw box file
}
\examples{
filePath = "/ds/raw_data_kemp/FIELD/Westerholz/SNB/RAWDATA/2017/2017.01.31/184/BOX001.TXT"
filePath = "/ds/raw_data_kemp/FIELD/Westerholz/SNB/RAWDATA/2017/2017.03.14/83/BOX001.TXT"
				 x = diagnose_raw_txt(filePath); x; x
}
\seealso{
\code{\link{diagnose_raw_dir}}
}
\author{
MV
}
