% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_rawdata_checks.R
\name{diagnose_raw_txt_v2}
\alias{diagnose_raw_txt_v2}
\title{Output diagnostics on raw SNB files (v2)}
\usage{
diagnose_raw_txt_v2(filePath)
}
\value{
A data.table
}
\description{
Output diagnostics of a raw box file
}
\examples{
    
\dontrun{
filePath = "/ds/raw_data_kemp/FIELD/Westerholz/SNB/RAWDATA_v2/2017/2017.03.03/122/BOX0122.TXT"
	}			 x = diagnose_raw_txt_v2(filePath); x; x
}
\author{
MV
}
