% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_admin_rawtxt_to_db_PIPELINE.R
\name{drop_and_reload}
\alias{drop_and_reload}
\title{Unload and reload files}
\usage{
drop_and_reload(con, id, ...)
}
\arguments{
\item{con}{a connection object}

\item{id}{id}

\item{...}{goes to  scidb_snbUpdater}
}
\description{
Unload and reload files by ID
}
