% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ELEMENTS_ds_to_scidb.R
\name{drop_by_id}
\alias{drop_by_id}
\title{drop ALL data associated with a given ID}
\usage{
drop_by_id(con, id, db = getOption("snbDB"))
}
\arguments{
\item{con}{a connection object}

\item{id}{the \code{id}-s in file status and the associated box tables.}
}
\description{
The function removes all lines in both file_status and box tables associated with a given id.
}
\author{
MV
}
