% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fast_POSIXct}
\alias{fast_POSIXct}
\title{fast_POSIXct}
\usage{
fast_POSIXct(x, tz = "CEST")
}
\arguments{
\item{x}{timestamps (YYYY-mm-dd HH:MM:SS)}

\item{tz}{local timezone}
}
\value{
POSIXct vector
}
\description{
Quickly converts local timestamps to a POSIXct vector
}
