% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_plot_functions.R
\name{feedingPlot}
\alias{feedingPlot}
\title{Function to plot approaches to/duration of time spent inside the nestbox.}
\usage{
feedingPlot(x, variable = "rate", dayTo = "dayToHatch", maleCol = "blue",
  midCol = "white", femaleCol = "red", showSexDifference = TRUE,
  QUANTILE = FALSE)
}
\arguments{
\item{x}{A data.table as provided by the functions by the function breedingStage,
to which the output of the function "feeding" in put in.}

\item{variable}{"rate" or "time_spent". Should the approach rate ("rate") or the duration inseide the nestbox
(time_spent) be plotted? Defaults to "rate".}

\item{dayTo}{dayToHatch or dayToFristEgg. In relation to which event should the data be plotted?
Defaults to "dayToHatch".
Note that to use real dates, a dayToHatch column with a numerical date object can be provided.}

\item{maleCol}{Male colour. Defaults to "blue".}

\item{midCol}{Baseline colour (no activity/similar male and female activity). Defaults to "white".}

\item{femaleCol}{Female colour. Defaults to "red".}

\item{showSexDifference}{Logical. Should differences between male and female behaviour be plotted?
Defaults to TRUE.}

\item{QUANTILE}{Logical. Should absolute values be plotted, or should the colour be quantile-based?
The advantage of using quantiles is that in the presence of outliers, the plot can be meaningful.
Defaults to FALSE.}
}
\value{
A ggplot graphic.
}
\note{
This function uses the sex column to specify the identity of the bird. This has two consequences:
\itemize{
\item If only the resident pair should be plotted, subset the data.table x accordingly before plotting.
\item The sex-column can be (manually) specified as needed, i.e. if events from NA transponder readings should be plotted
as well.
}
}
\examples{
con = dbcon(user = "YOUR_USER_NAME", host = getOption("host") )
x = queryEvents(con = con, box = 256, from = "2014-04-05", to = "2014-05-25", includeFIELD = TRUE)
x = queryResidents(con = con, x = x)
x = feeding(x = x, timeScale = 10, from = '2014-04-05', to = '2014-05-25')
x = breedingStage(con = con, x = x)
closeCon(con)

#Quantiles
feedingPlot(x, QUANTILE = TRUE)
feedingPlot(x, QUANTILE = FALSE)

#Sex difference
feedingPlot(x, showSexDifference = TRUE)
feedingPlot(x, showSexDifference = FALSE)

#seasonal timing
feedingPlot(x, dayTo = 'dayToHatch')
feedingPlot(x, dayTo = 'dayToFirstEgg')

#feeding rate vs. duration in nestbox.
feedingPlot(x, variable = 'rate')
feedingPlot(x, variable = 'time_spent') #note the bug on ca. day 25 in the male figure: overnight sleep

#adjust colours
feedingPlot(x, maleCol = 'black', midCol = 'white', femaleCol = 'black', showSexDifference = FALSE)
}
\author{
LS
}
