% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_EVENTS_elements_v1.R
\name{fetch_data_v1}
\alias{fetch_data_v1}
\title{Fetch data from database}
\usage{
fetch_data_v1(con, box = NA, from = NA, to = NA)
}
\arguments{
\item{con}{An open connection}

\item{box}{The box number from which to collect data}

\item{from}{Defines the starting point at which to fetch data. This can refer either to the r_pk column (numeric)
or can be a date and time (in the format "YYYY-MM-DD HH:MM:SS")}

\item{to}{Defines the end point to which to fetch data. This can refer either to the r_pk column (numeric)
or can be a date and time (in the format "YYYY-MM-DD HH:MM:SS")}
}
\description{
This function fetches the raw data from the database using either datetime_ or r_pk as a reference column.
}
\examples{
see function events
}
\seealso{
\code{\link{events}}, \code{\link{remove_faulty_data}}, \code{\link{trim_data}},
\code{\link{fetch_ins_outs}}, \code{\link{assign_direction}}, \code{\link{concat_events}},
}
\author{
LS
}
