% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_EVENTS_elements.r
\name{fetch_data_v2}
\alias{fetch_data_v2}
\title{Fetch data from databases}
\usage{
fetch_data_v2(con, box = NA, from = NA, to = NA)
}
\arguments{
\item{con}{A connection as established e.g. by dbcon()}

\item{box}{The box number for which the data should be fetched.}

\item{from}{From which point onwards should the data be fetched? Can be a date-time object or an r_pk (column r_pk from the bxxx tables)}

\item{to}{Until which point should the data be fetched? Can be a date-time object or an r_pk (column r_pk from the bxxx tables)}
}
\value{
A data.table with all the information from the bxxx table, and an additional box column. Any rows where the column datetime_ is NA are removed.
}
\description{
Fetches data from the bxxx databases.
}
\examples{
#Not run
#see help of function events_v2
?events_v2()
}
