% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_EVENTS_elements_v1.R
\name{fetch_ins_outs_v1}
\alias{fetch_ins_outs_v1}
\title{Split data into different pieces of activity for SNB data version v1}
\usage{
fetch_ins_outs_v1(x, time_threshold)
}
\arguments{
\item{time_threshold}{Time in seconds. If the time difference between two consecutive lines is larger or equal to time_threshold, the rows will be treated as independent pieces of activity. Defaults to 2.}
}
\description{
This function fetches individual in and out events based on a time threshold and transponder readings.
If two transponder readings occur close in time, the rows in between are removed, because it is unclear to which
individual this data belongs, and the information about the direction of the movement is marked as "unclear".
}
\details{
See \link{events_v1} for further information.
}
\examples{
#Not run
#see help of function events_v1
?events_v1()
}
