% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_EVENTS_elements_v2.r
\name{fetch_ins_outs_v2}
\alias{fetch_ins_outs_v2}
\title{Split data into different pieces of activity}
\usage{
fetch_ins_outs_v2(x, tr_threshold = 5, broken_LB_threshold = 600,
  cluster_events_threshold = 1)
}
\arguments{
\item{x}{A data.table as supplied by \link{fetch_data_v2}.}

\item{tr_threshold}{Time in seconds within which the light barriers are assigned to a specific transponder}

\item{broken_LB_threshold}{Time in seconds after which a light barrier that is constantly on will be marked as off.}

\item{cluster_events_threshold}{Time in seconds within which different oieces of activity will be considered as belonging together.}
}
\description{
Split data into different pieces of activity, e.g. one "IN" and one "OUT"
}
\details{
See \link{events_v2} for further information.
}
\examples{
#Not run
#see help of function events_v2
?events_v2()
}
