% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{force.tz}
\alias{force.tz}
\title{force timezones.}
\usage{
force.tz(x, new.tz, diff.tz = NULL)
}
\arguments{
\item{x}{A POSIXct object}

\item{new.tz}{time zone to convert to}

\item{diff.tz}{Optional. The difference between the current and the new timezone, if known. Note that providing this variable will increase the speed of the function 100-fold. Defaults to NULL, in which case the difference is calculated.}
}
\description{
force timezones.
}
\details{
Note if diff.tz is not supplied, this is time-consuming because it uses substring (ca. 1.5 seconds/100000 items). However, if diff.tz is not supplied should handle CET/CEST and other time-zone mixes correctly.
}
\note{
Note that the time zones that are defined by its offset to UTC are coded e.g. "Etc/GMT-2", where the sign of the code is opposite to the expectation: "Etc/GMT-2" refers to the time zone GMT+2 (which is also printed).
}
