% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ELEMENTS_events.R
\name{helper_fun_direction}
\alias{helper_fun_direction}
\title{fetch LB direction using the rule or 'decision tree'; this function underlies the function "assign direction"}
\usage{
helper_fun_direction(x, rule)
}
\arguments{
\item{x}{a vector of light barriers}

\item{rule}{a decision rule}
}
\value{
a vector with directions
}
\description{
fetch LB direction using the rule or 'decision tree'; this function underlies the function "assign direction"
}
\examples{
internal function, not to be used otherwise.
}
\seealso{
\code{\link{assign_direction}}
}
\author{
LS
}
