% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ELEMENTS_events.R
\name{helper_shift_compare}
\alias{helper_shift_compare}
\title{helper function to compare columns that are shifted by 1}
\usage{
helper_shift_compare(x, x_1 = NULL, FUN = function(x, y)
  return(ifelse(!is.na(x) & !is.na(y), x == y, NA)), move = "up",
  mode = "logical", missing_value = NA)
}
\arguments{
\item{x}{a vector}

\item{x_1}{a second vector which is compared to x. If not supplied, x is recycled by being moved up or down (as specified by "move")}

\item{FUN}{a function used for comparing x and x_1}

\item{mode}{mode of the output variable}

\item{missing_value}{value inserted for missing values}

\item{"up"}{or "down": move should x_1 (or x) be moved up or down?}
}
\value{
a vector
}
\description{
helper function to compare columns that are shifted by 1
}
\examples{
helper_shift_compare(rep(1:5, each = 2))

}
\author{
LS
}
