% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_rawtxt_to_db_ELEMENTS.R
\name{hot_files}
\alias{hot_files}
\title{Hot txt files}
\usage{
hot_files(con, p = getOption("path.to.raw_v2"), db = getOption("snbDB_v2"))
}
\arguments{
\item{con}{A connection object}
}
\value{
\code{data.table}
}
\description{
Files listed in \code{file_status} table (upload_status is NULL) which are ready to process and upload
}
\author{
MV
}
