% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fileop.R
\name{hwid}
\alias{hwid}
\title{Extract hardware ID.}
\usage{
hwid(p, n = 300)
}
\arguments{
\item{p}{path to the snb file}
}
\value{
a  character string (length 1).
}
\description{
Extract hardware ID from a TXT raw data file.
}
\examples{
\dontrun{
hwid(p =  "/ds/raw_data_kemp/FIELD/Westerholz/SNB/RAWDATA_v2/2016/2016.06.07/267/BOX0267.TXT")
}
}
\author{
MV
}
