% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_demo_systems.R
\name{install_demo_SNB}
\alias{install_demo_SNB}
\title{Installs a demo system on 127.0.0.1}
\usage{
install_demo_SNB(user = "testuser", host = "127.0.0.1",
  rawdata_root = paste(tempdir(), "SNB_demo_RAWDATA", sep = "/"),
  db = "SNB_demo")
}
\description{
Installs a demo system on 127.0.0.1, prepares a temp RAWDATA folder and changes the options
}
\examples{
\dontrun{
install_demo_SNB()
}

}
