% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_demo_system}
\alias{install_demo_system}
\title{Installs a demo system on 127.0.0.1}
\usage{
install_demo_system(user, host = "127.0.0.1",
  rawdata_root = paste(tempdir(), "snb_demo_RAWDATA", sep = "/"),
  db = "demo")
}
\description{
Installs a demo system on 127.0.0.1, prepares a temp RAWDATA folder and changes the options
}
\examples{
\dontrun{
install_demo_system('mihai', '127.0.0.1')
}

}
