% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int2b}
\alias{int2b}
\title{integer to box name}
\usage{
int2b(x)
}
\arguments{
\item{x}{box names (numeric)}
}
\value{
character vector
}
\description{
convert an integer as b00x
}
\examples{
int2b(1)
}
\author{
MV
}
