% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_eventsDataQueries.R
\name{ivi}
\alias{ivi}
\title{Calculate intervisit intervals}
\usage{
ivi(x, includeNA = FALSE, residentOnly = TRUE)
}
\arguments{
\item{x}{A data table as provided by query events}

\item{includeNA}{Should visits without a transponder identity be included?}

\item{residentOnly}{Should only residents be included?}
}
\value{
A data.frame with one additional column, namely the intervisit intervals.
}
\examples{
con = dbcon(user = "YOUR_USER_NAME", host = getOption("host") )
x = queryEvents(con = con, box = 26, from = "2014-05-17", to = "2014-05-18", includeFIELD = TRUE)
x = queryResidents(con = con, x = x)
closeCon(con)
x = ivi(x)
plot(density(na.omit(x[, ivi])))
}
\author{
LS
}
