% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_names_local}
\alias{make_names_local}
\title{make names of data.table objects local in order to suppress warnings during package compilation.}
\usage{
make_names_local(data)
}
\arguments{
\item{x}{A data table}
}
\description{
make names of data.table objects local in order to suppress warnings during package compilation.
}
\details{
use in the beginning of every function that calls a data.table object and its columns.
}
