% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_queries.R
\name{overnight}
\alias{overnight}
\title{overnight}
\usage{
overnight(username = getOption("DB_user"), host = getOption("host"),
  buffer = 1, date = Sys.Date() - 1)
}
\arguments{
\item{username}{username}

\item{host}{host}

\item{buffer}{(hours)   sunrise + buffer}

\item{date}{date, default to last date in file_status table}
}
\description{
overnight
}
\examples{
overnight(buffer = 2)

}
