% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path2box}
\alias{path2box}
\title{box from file path (v1)}
\usage{
path2box(x)
}
\description{
extract box from a file path string
}
\examples{
 path2box("/ds/raw_data_kemp/FIELD/Westerholz/SNB_v2/RAWDATA/2016/2016.01.26/011/BOX001.TXT")
}
\author{
MV
}
