% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path2year}
\alias{path2year}
\title{year from file path}
\usage{
path2year(x)
}
\description{
extract year from a file path string
}
\examples{
 path2year("/ds/raw_data_kemp/FIELD/Westerholz/SNB/RAWDATA/2016/2016.01.26/004/BOX001.TXT")
}
\author{
MV
}
