% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_field_functions.R
\name{plot_from_raw}
\alias{plot_from_raw}
\title{Plot raw data in the field}
\usage{
plot_from_raw(path = "/home/bt/Desktop/", box = 999)
}
\arguments{
\item{path}{The path to the BOX001.TXT file. Defaults to the desktop (/home/bt/Desktop/)}

\item{box}{A box number (defaults to 999)}
}
\value{
An actogram (.pdf) that is located at the specified path and a data table with the underlying data.
}
\description{
Plot raw data in the field
}
\examples{
#copy a BOX001.TXT to the desktop if the blue tit user and run
x = plot_from_raw()
x
#have a look at your desktop. A file called box_999.pdf should have appeared.
}
