% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_eventsDataQueries.R
\name{posthoc_combineEvents}
\alias{posthoc_combineEvents}
\title{Function that combines events from the EVENTS table based on used-specified rules.}
\usage{
posthoc_combineEvents(x, compareWith = c("previous", "next")[1],
  assignNA = TRUE, timeCutoff = 3000, jValidity = c("OUT", "PASS",
  "BACK"), compareValidity = c("IN"))
}
\arguments{
\item{x}{A data.table as returned by queryEvents().}

\item{compareWith}{Is the rule based on a comparison with a previous or with the next line?
Can be either "previous" or "next".}

\item{assignNA}{Logical. Should events with NA transponder readings be assigned to visits 
with a transponder reading, if the specified rules are fulfilled? Defaults to FALSE.}

\item{timeCutoff}{A time limit (in seconds) up onto which events can be assigned together. For example,
a value of 30 means that any events that are more than 30 seconds apart will not be grouped together.}

\item{jValidity}{The validity assignment of the focal row.}

\item{compareValidity}{The validity assignement of the row to compare with.}
}
\value{
A data.table of the same structure as queryEvents(), but with the lines concatenated according to the
specified rule.
}
\description{
This function can be used to fine-tune the assignment performed in the events-function based on
the specific needs of the user.
}
\details{
The comparison is performed within a specific transponder number (or includes NAs, if specified with assign NA).
The example is based on a rule used by Peter Santema et al. (ca. 2015) for
 the experiment on parental feeding and offspring begging. It has therefore been checked carefully for its validity.
}
\note{
If compareWith == 'previous', then the progress bar will run backwards, because the data.table is
run through backwards. 
Note that the events function works within a file (within column "id"). This means that ins and outs that come from 
separate files will not automatically be grouped together. This function runs across files (see example).
}
\examples{
con = dbcon(user = "YOUR_USER_NAME", host = getOption("host") )
x = queryEvents(con = con, box = 254, from = "2015-05-20 07:49:00", to = "2015-05-20 07:55:00", includeFIELD = TRUE)
setkey(x, id, e_pk) #setkey is used to sort the data.table for visualization; it does not need to be included into other scripts.
x2 = posthoc_combineEvents(x)
x
x2
#Note that this example was part of an experiment where the transponder was entered before and after the SD-card was pulled. Apparently 
the bird left the nestbox in between these two test-transponder events.
}
\author{
LS
}
