% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_eventsDataQueries.R
\name{queryEvents}
\alias{queryEvents}
\title{Fetch data from events and add ID and sex}
\usage{
queryEvents(con, box = NA, from = NA, to = NA, includeFIELD = FALSE)
}
\arguments{
\item{box}{The box from which to fetch data. If not supplied, all data is imported.}

\item{from}{Character. Date (YYYY-MM-DD) or date and time (YYYY-MM-DD HH:MM:SS) from which to import data.
If not supplied all data are used (not recommended due to computing time issues).}

\item{to}{Character. Date (YYYY-MM-DD) or date and time (YYYY-MM-DD HH:MM:SS) up to which to import data.
If not supplied all data are used.}

\item{includeFIELD}{Logical. Whether or not to include FIELD-database-based IDs. Be aware that these IDs 
are not cross-checked at that stage. By default only data from BTatWESTERHOLZ will be used.}

\item{user}{Your username, as specified in "saveCredentials()"}

\item{host}{Hostname. Defaults to scidb.orn.mpg.de}
}
\value{
A data.table object.
}
\note{
From is included (>=), to is excluded (<).
}
\examples{
con = dbcon(user = "YOUR_USER_NAME", host = getOption("host") )
x = queryEvents(con = con, box = 26, from = "2014-05-17", to = "2014-05-18", includeFIELD = TRUE)
closeCon(con)
}
\seealso{
\code{\link{queryResidents}}, \code{\link{feeding}}, \code{\link{feedingPlot}}
\code{\link{actogramPlot}}, \code{\link{events}}
}
\author{
LS
}
