% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_eventsDataQueries.R
\name{queryResidents}
\alias{queryResidents}
\title{Add residency to events imported by the "queryEvents" function.}
\usage{
queryResidents(con, x)
}
\arguments{
\item{x}{A data.table as returned by the function "queryEvents"}

\item{user}{Your username, as specified in "saveCredentials()"}

\item{host}{Hostname. Defaults to scidb.orn.mpg.de}
}
\value{
A data.table object.
}
\examples{
con = dbcon(user = "YOUR_USER_NAME", host = getOption("host") )
x = queryEvents(con = con, box = 26, from = "2014-05-17", to = "2014-05-18", includeFIELD = TRUE)
x = queryResidents(con = con, x = x)
closeCon(con)
}
\seealso{
\code{\link{queryEvents}}, \code{\link{feeding}}
}
\author{
LS
}
