% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rawstring2datetime}
\alias{rawstring2datetime}
\title{snb string to datetime_}
\usage{
rawstring2datetime(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector yyyy-dd-mm hh:mm:ss
}
\description{
convert an SNB string to a datetime object
}
\examples{
 rawstring2datetime('0212151805550010') # old format
}
\author{
MV
}
