% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_butler_ELEMENTS.R
\name{read.boxnumber}
\alias{read.boxnumber}
\title{looks up the box number}
\usage{
read.boxnumber(path, boxid, hwidCheck = TRUE)
}
\arguments{
\item{path}{file path}

\item{boxid}{data.table with current box - hwid}

\item{hwidCheck}{check hwid}
}
\value{
data.table containing Boxnummer.txt info
}
\description{
looks up the box number in BoxNNNN.txt and optionally cross-checks it with the HW-ID
}
\examples{
 require(sdb)
 boxid <- dbq( q= 'select box, hwid from boxid order by box, datetime_ desc',
         db = getOption('snbDB_v2'), host = getOption('host'), user =getOption('DB_user'))[!duplicated(box)]
x = cardReader()
z = x[, read.boxnumber(mountpoint, boxid = boxid), by = .(row.names(x)) ]
}
\author{
MV
}
