% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_queries.R
\name{read.fstatus}
\alias{read.fstatus}
\title{read file status}
\usage{
read.fstatus(con, date)
}
\arguments{
\item{con}{a db connection}
}
\value{
a data.table; authors tests transponders are returned as well
}
\description{
read a subset of the the file status  given a date
}
\examples{
    con = dbcon('mihai', host = 'scidb.mpio.orn.mpg.de')
read.fstatus(con, "2016.02.17")
}
\author{
MV
}
