% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_rawdata_checks.R
\name{read_raw_transponders}
\alias{read_raw_transponders}
\title{Read transponders}
\usage{
read_raw_transponders(date, outDirLocation = getOption("path.to.raw"))
}
\arguments{
\item{date}{a character vector formated as 'yyyy.mm.dd' (same as the directory date holding the raw data)}

\item{outDirLocation}{the location of the raw data; the default is getOption('path.to.raw')}
}
\value{
a data.table
}
\description{
Read all lines containing transponders
}
\examples{
x = read_raw_transponders("2016.05.10")
check RTC drift
con = dbcon('mihai', host = 'scidb.mpio.orn.mpg.de')
}
\author{
MV
}
