% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_SF.R
\name{read_sf_dir}
\alias{read_sf_dir}
\title{read SF directory}
\usage{
read_sf_dir(D, nextid = NA, basepath = getOption("path.to.raw_v2_SF"))
}
\arguments{
\item{D}{path to directory}

\item{nextid}{next file ID in the db table.}
}
\value{
a  \code{data.table} .
}
\description{
read all SF files in a given root directory
}
\examples{
\dontrun{
read_sf_dir('/ds/raw_data_kemp/FIELD/Westerholz/SF/RAWDATA/2018/')
}
}
\author{
MV
}
