% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_SF.R
\name{read_sf_txt}
\alias{read_sf_txt}
\title{read Smart Feeder files}
\usage{
read_sf_txt(f)
}
\arguments{
\item{f}{path to the SF file}
}
\value{
a  \code{data.table} .
}
\description{
read one SF file as data.table
}
\examples{
x = read_sf_txt(f = system.file(package = 'SNB2', 'sample_files', 'BOX1020.TXT'))
}
\author{
MV
}
