% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_SF.R
\name{read_sf_txt_v2}
\alias{read_sf_txt_v2}
\title{read SF data}
\usage{
read_sf_txt_v2(f)
}
\arguments{
\item{f}{path to the SF file}
}
\value{
a  \code{data.table} .
}
\description{
read an SF file as data.table
}
\examples{
ex_file = system.file(package = 'SNB2', 'sample_files', 'BOX1020.TXT')
#' x = read_sf_txt_v2(f = ex_file)

}
\author{
MV
}
