% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_v2.R
\name{read_snb_txt_v2}
\alias{read_snb_txt_v2}
\title{read SNB data}
\usage{
read_snb_txt_v2(filePath)
}
\arguments{
\item{filePath}{path to the snb file}
}
\value{
a  \code{data.table} .
}
\description{
read an SNB file as data.table-s (only transponder and Light barrier data are parsed).
}
\examples{
x = read_snb_txt_v2(filePath = "/ds/raw_data_kemp/FIELD/Westerholz/SNB/RAWDATA_v2/2016/2016.06.07/267/BOX0267.TXT")

}
\author{
MV
}
