% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readRaw}
\alias{readRaw}
\title{Read raw}
\usage{
readRaw(filePath)
}
\arguments{
\item{filePath}{file path to the box file}
}
\value{
a \code{data.table} with two columns: V (the data string )  and bout_length (N time the string is repeated)
}
\description{
Read raw box txt file without conversion and string splitting
}
\examples{
readRaw('/ds/raw_data_kemp/FIELD/Westerholz/SNB/RAWDATA/2016/2016.01.26/004/BOX001.TXT')
# without embedded null string
readRaw("/ds/raw_data_kemp/FIELD/Westerholz/SF/RAWDATA/2014/2014.11.20/014/BOX001.TXT")
}
\author{
MV
}
