% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_fileop.R
\name{readRaw_v2}
\alias{readRaw_v2}
\title{Read raw}
\usage{
readRaw_v2(filePath)
}
\arguments{
\item{filePath}{file path to the box file}
}
\value{
a \code{data.table}
}
\description{
Read raw box txt file without any conversion and string splitting
}
\examples{
\dontrun{
a = readRaw_v2(filePath = 
        "/ds/raw_data_kemp/FIELD/Westerholz/breeding_2016_snb_heaven/07062016/122/BOX0122.TXT")
}
}
\author{
MV
}
