% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ELEMENTS_events.R
\name{remove_faulty_data}
\alias{remove_faulty_data}
\title{Remove any data that cannot be interpreted}
\usage{
remove_faulty_data(x)
}
\arguments{
\item{x}{a data.frame as provided by the function fetch_data}

\item{con}{a connection}
}
\value{
a data.frame where misspecified data is missing
}
\description{
Remove any data that cannot be interpreted
}
\examples{
see function events
}
\seealso{
\code{\link{events}}, \code{\link{fetch_data}}, \code{\link{trim_data}},
\code{\link{fetch_ins_outs}}, \code{\link{assign_direction}}, \code{\link{concat_events}},
}
\author{
LS
}
