% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_TXT_to_b000.R
\name{scidb_snbUpdater.b000}
\alias{scidb_snbUpdater.b000}
\title{txt Box files (txt) to db b000 tables pipeline}
\usage{
scidb_snbUpdater.b000()
}
\value{
TRUE if all txt are updated
}
\examples{
\dontrun{
 require(SNB2)
 
 demo_setup(install.test.db = TRUE, admin.user = 'mihai')
 scidb_snbUpdater.b000 ()
}

}
\author{
MV
}
