% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_b000_to_transponders.R
\name{scidb_snbUpdater.transponders}
\alias{scidb_snbUpdater.transponders}
\title{b000 tables to transponders table pipeline}
\usage{
scidb_snbUpdater.transponders(file = "~/scidb_snbUpdater.log")
}
\value{
TRUE on success
}
\description{
b000 tables to transponders table pipeline
}
\examples{
\dontrun{
 require(SNB2)
 
 demo_setup(install.test.db = TRUE, admin.user = 'mihai')
 scidb_snbUpdater.b000()
 scidb_snbUpdater.transponders()
}

}
\author{
MV
}
