% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_butler_ELEMENTS.R
\name{sdcard_format}
\alias{sdcard_format}
\title{Format SD cards}
\usage{
sdcard_format(x = card_format_status(), ids, password, format = FALSE)
}
\arguments{
\item{x}{card_format_status() output}

\item{password}{password}
}
\author{
MV
}
