% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_butler_ELEMENTS.R
\name{sdcard_prepare}
\alias{sdcard_prepare}
\title{Prepare SD cards}
\usage{
sdcard_prepare(x = card_prepare_status(), ids)
}
\arguments{
\item{x}{card_prepare_status() output}

\item{password}{password}
}
\author{
MV
}
