% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_eventsDataQueries.R
\name{sleep}
\alias{sleep}
\title{Find sleeping birds}
\usage{
sleep(x, na.only = TRUE, output = c("data", "box")[2])
}
\arguments{
\item{x}{A data.table as supplied by queryEvents()}

\item{na.only}{Should only unknown birds (NA transponder reading) or all birds be displayed?}

\item{output}{Should only box numbers or the entire dataset be returned?}
}
\value{
A vector of box numbers or a data.table (as specified bz the parameter "output")
}
\description{
For use for winter nightcatches.
}
\note{
It max be worth double-checking the returned nestboxes using the actogramPlot function (see example)
}
\examples{
#Run for one nestbox only 
con = dbcon(user = "YOUR_USER_NAME", host = getOption("host") )
x = queryEvents(con = con, box = 26, from = "2014-04-17", to = "2014-05-18", includeFIELD = TRUE)


#Run function for multiple nestboxes
input = data.table(box = 1:277, from = "2014-04-17", to = "2014-04-18") #this may take a minute to run

output = list()
bar  = txtProgressBar(min = 1, max = nrow(input), initial = 1, style = 3, width = floor(getOption("width")/2)  )
for(i in 1 : nrow(input)) {
 setTxtProgressBar(bar, i)
  x = queryEvents(con = con, box = input[i,box], from = input[i,from], to = input[i,to], includeFIELD = TRUE); if(nrow(x) == 0) next;
  x -> output[[length(output)+1]]
}
output = rbindlist(output)
closeCon(con)

S1 = sleep(output)

#plot first returned nestbox
actogramPlot(subset(output, box == S1[1]))

#other parameter settings
S2 = sleep(output, na.only = FALSE)
S3 = sleep(output, na.only = FALSE, output = 'data')


}
\author{
LS
}
