% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{package}
\name{SNB}
\alias{SNB}
\alias{.onLoad}
\alias{SNB-package}
\title{Smart Box Data management}
\usage{
Overview help files can be found at http://scidb.mpio.orn.mpg.de
}
\description{
An all-in-one package for data management, user-level functions and hardware feedback.
}
