% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SNB2}
\alias{SNB2}
\alias{.onLoad}
\title{Smart Box Data management}
\usage{
.onLoad(libname, pkgname)
}
\description{
An all-in-one package for data management, user-level functions and hardware feedback.
}
\note{
Overview help files can be found at http://scidb.mpio.orn.mpg.de
}
