% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snbstring2date_v2}
\alias{snbstring2date_v2}
\title{Extract date}
\usage{
snbstring2date_v2(x)
}
\arguments{
\item{x}{character vector(s)}
}
\value{
\code{POSIXct} object
}
\description{
Extract date from raw data string
}
\examples{
   
snbstring2date_v2(x = '20170418-171742.202 Transponder: 4B76C4B43A6F0001')
snbstring2date_v2(x = '20170418-171742 Transponder: 4B76C4B43A6F0001')
}
\author{
MV
}
