% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_butler_ELEMENTS.R
\name{system_status}
\alias{system_status}
\title{system screening: looks up all the relevant system settings}
\usage{
system_status(pulldate = Sys.Date(), password = "")
}
\arguments{
\item{pulldate}{Sd card pool date}

\item{password}{password}
}
\value{
data.table with 3 columns
}
\examples{
head(system_status(password = 'some char'))
}
\author{
MV
}
