% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_queries.R
\name{tetr}
\alias{tetr}
\title{test transponders}
\usage{
tetr(condb)
}
\arguments{
\item{condb}{a db connection}
}
\value{
a data.table
}
\description{
the latest test transponder and their holders (authors)
}
\examples{
 
\dontrun{
con = dbcon('snbAdmin', host = 'scidb.mpio.orn.mpg.de')
tetr(con)
}
}
\author{
MV
}
