% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/A_EVENTS_elements_v2.r
=======
% Please edit documentation in R/A_EVENTS_elements.r
>>>>>>> 627070551dd4ed364ab330fe54c7d4dcdc78d8e3
\name{translate_validity_v2}
\alias{translate_validity_v2}
\title{Simplify direction of activity}
\usage{
translate_validity_v2(x)
}
\arguments{
\item{x}{A data.table as supplied by \link{combine_front_v2}.}
}
\description{
Simplify the information of the direction: Translate the detailed directional information into something radable.
}
\details{
See \link{events_v2} for further information.
}
\examples{
#Not run
#see help of function events_v2
?events_v2()
}
