% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ELEMENTS_events.R
\name{trim_data}
\alias{trim_data}
\title{Remove meaningless light barriers}
\usage{
trim_data(x)
}
\arguments{
\item{x}{a data.frame as provided by the function remove_faulty data}
}
\value{
a data.frame where "meaningless" light barriers are removed
}
\description{
This function removes the light barriers 00, 02, or 10, if not preceded or followed by a 12, because these lines do not infer information regarding the direction. Lines with transponder readings are not deleted.
}
\examples{
see function events
}
\seealso{
\code{\link{events}}, \code{\link{fetch_data}}, \code{\link{remove_faulty_data}},
\code{\link{fetch_ins_outs}},\code{\link{assign_direction}},\code{\link{concat_events}},
}
\author{
LS
}
