% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ELEMENTS_ds_to_scidb.R
\name{update_bTables}
\alias{update_bTables}
\title{Update  box and file_status tables}
\usage{
update_bTables(con, dat, ui = FALSE, demo = FALSE,
  db = getOption("snbDB"))
}
\arguments{
\item{con}{a connection object}

\item{dat}{a list of \code{data.table}-s with attributes see \code{\link{load_clean_txt}}.}
}
\value{
the number of uploaded files
}
\description{
Update  box tables with parsed txt data (e.g. by \code{\link{load_clean_txt}}  )
}
\author{
MV
}
