% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_PIPELINE_events.R
\name{upload_events}
\alias{upload_events}
\title{Upload/update EVENTS table.}
\usage{
upload_events(con, id = NULL)
}
\arguments{
\item{con}{A connection to the database
id Defaults to NULL, can be a number or a vector. In this case the function will look for
new data by comparing the maximum r_pk from the bxxx tables with
the data in the events table. If `id` id specified, the corresponding
files will be processed from the bxxx tables to the EVENTS table.}
}
\value{
A progress bar.
}
\description{
Function that uploads new raw data to the EVENTS table.
}
\details{
This function relies on parallel processing.
}
\note{
\itemize{
\item 1. This function will only upload data where dates and times
are likely to be correct (based on the DATA_QUALITY table)
\item 2. This function has not been tested on a Windows computer and
is likely not to run on Windows
}
}
\examples{
#NOT RUN
con = dbcon(user = "snbAdmin", host = getOption("host"))
upload_events()
closeCon(con)
}
\author{
LS
}
