% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_VideoVsEvents.r
\name{videoEvaluation_v2}
\alias{videoEvaluation_v2}
\title{Load and assemble much of the data for video validation}
\usage{
videoEvaluation_v2(...)
}
\description{
Load and assemble much of the data for video validation
}
\details{
Called by \link{SNBvsVid_v2}, please follow the link for further details.
}
