% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_VideoVsEvents.r
\name{videoTimeCheck_v2}
\alias{videoTimeCheck_v2}
\title{Prepare the video data for comparison with SNB data by unit of time.}
\usage{
videoTimeCheck_v2(VideoData)
}
\description{
Prepare the video data for comparison with SNB data by unit of time.
}
\details{
Called by \link{SNBvsVid_v2}, please follow the link for further details.
}
