% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_rawtxt_to_db_ELEMENTS.R
\name{well_formated_directory}
\alias{well_formated_directory}
\title{Check the raw data directory}
\usage{
well_formated_directory(p = getOption("path.to.raw_v2"),
  y = year(Sys.Date()))
}
\arguments{
\item{p}{path to raw data, default to getOption("path.to.raw")}
}
\value{
logical
}
\description{
The function checks the proper directory format: \code{ year/date[prefix]/box/b00x.txt}
}
\examples{
well_formated_directory(getOption("path.to.raw_v2"))

}
\author{
MV
}
