% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_box_activity.R
\name{box_activity}
\alias{box_activity}
\title{Box activity}
\usage{
box_activity(lastday, days_before = 1)
}
\arguments{
\item{lastday}{a Date object}
}
\value{
A data.table
}
\description{
Box activity estimate based on light-barrier to transponder ratio
}
\examples{
    
\dontrun{
   x = box_activity()
   require(ggplot2)   
   require(ggrepel)   
  
  ggplot(x, aes(x = log(total_tr_reads+1), y = log(total_lb_reads), color = `N_transp\ndetected`  ) ) +
  geom_point() + geom_abline(intercept = 0, slope = 1) + 
  geom_text_repel(data =x[unk_act > .5], 
   aes(x = log(total_tr_reads+1), y = log(total_lb_reads), label = box) ) +
   xlab('log [ total transponder reads]') + 
   ylab('log [ total light-barrier reads]') + 
    scale_colour_brewer(palette = "Set1")

 }            
}
\author{
MV
}
