% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_butler_ELEMENTS.R
\name{cardReader}
\alias{cardReader}
\title{Finds attached card reader devices (a wrapper to lsblk)}
\usage{
cardReader(cardsize = getOption("cardSize"))
}
\arguments{
\item{cardsize}{default to getOption('cardSize')}
}
\value{
data.table containing info on HOTPLUG,MOUNTPOINT,KNAME,MODEL,VENDOR,TRAN,FSTYPE
}
\description{
Finds attached card reader devices (a wrapper to lsblk)
}
\examples{
cardReader()
}
\author{
MV
}
