% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fileop.R
\name{data_dirs}
\alias{data_dirs}
\title{List all data directories}
\usage{
data_dirs(p = getOption("path.to.raw_v2"))
}
\arguments{
\item{p}{path to raw data, default to getOption("path.to.raw")}
}
\value{
a data.table
}
\author{
MV
}
