% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_queries.R
\name{dbqSNB}
\alias{dbqSNB}
\title{query raw data, all b_ tables at once}
\usage{
dbqSNB(username, host, q = "SELECT * FROM boxtables limit 1",
  db = getOption("snbDB"), testtr_remove = TRUE, ncores = 4)
}
\arguments{
\item{username}{username}

\item{host}{host}

\item{q}{query,boxtables kw should be used instead of bnnn box name.}

\item{testtr_remove}{default to TRUE}
}
\description{
query raw data, all b_ tables at once
}
\examples{
# Simple select
dbqSNB('mihai', host = "scidb.mpio.orn.mpg.de", 'SELECT * FROM boxtables limit 1')

# Last entry
dbqSNB('mihai', host = "scidb.mpio.orn.mpg.de", 'SELECT * FROM boxtables ORDER BY r_pk desc limit 1')

# Select last n days
x = dbqSNB('mihai', host = "scidb.mpio.orn.mpg.de",
'SELECT datetime_, LB, transp FROM  boxtables
   WHERE datetime_ >= DATE_ADD( (select max(datetime_) from  boxtables) , INTERVAL -3 DAY) AND
   YEAR(datetime_) = YEAR(CURDATE() )
')

}
