% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_rawdata_checks_v2.R
\name{diagnose_pull_v2}
\alias{diagnose_pull_v2}
\title{Diagnose directory of one pull (i.e. a directory)}
\usage{
diagnose_pull_v2(date, outDirLocation = getOption("path.to.raw_v2"),
  shiny = FALSE)
}
\arguments{
\item{date}{a character vector formated as 'yyyy.mm.dd' (same as the directory date holding the raw data)}

\item{outDirLocation}{the location of the raw data; the default is getOption('path.to.raw_v2')}
}
\value{
a data.table
}
\description{
Output diagnostics of a directory containing raw box files
}
\examples{
x = diagnose_pull_v2(date = "2017.03.03")
}
\author{
MV
}
