% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_rawdata_checks.R
\name{diagnose_raw_txt_v2}
\alias{diagnose_raw_txt_v2}
\title{Output diagnostics on raw SNB files (v2)}
\usage{
diagnose_raw_txt_v2(filePath)
}
\value{
A data.table
}
\description{
Output diagnostics of a raw box file
}
\examples{
    
\dontrun{
options(host = 'scidb.mpio.orn.mpg.de') 
filePath = "/ds/raw_data_kemp/FIELD/Westerholz/SNB/RAWDATA_v2/2018/2018.04.03/1/BOX0001.TXT"
			 x = diagnose_raw_txt_v2(filePath); x; x
}
}
\author{
MV
}
