% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_PIPELINE_ds_to_scidb.R
\name{drop_and_reload}
\alias{drop_and_reload}
\title{Unload and reload files}
\usage{
drop_and_reload(con, id)
}
\arguments{
\item{con}{a connection object}

\item{id}{id}
}
\description{
Unload and reload files by ID
}
