% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_EVENTS_v1_v2_pipeline.r
\name{eva}
\alias{eva}
\title{Pipeline function to extract events}
\usage{
eva(username, df, setTZ = "Etc/GMT-2", time_threshold_v1 = 2,
  max_distance_v1 = 16 * 60 * 60, tr_threshold_v2 = 5,
  cluster_events_threshold_v2 = 2, max_distance_v2 = 16 * 60 * 60,
  cluster_fronts_threshold_v2 = 5, no_front_v2 = NULL)
}
\arguments{
\item{username}{the username that is supplied to dbcon for establishing the connection.}

\item{df}{a data.table that contains three columns about which data should be fetched: the box numbers, and the corresponding from and to (column names: box, from, to). See examples.}

\item{time_threshold_v1}{time_threshold argument for function \link{events_v1}}

\item{max_distance_v1}{max_distance argument for function \link{events_v1}}

\item{tr_threshold_v2}{tr_threshold argument for function \link{events_v2}}

\item{cluster_events_threshold_v2}{cluster_events_threshold argument for function \link{events_v2}}

\item{max_distance_v2}{max_distance argument for function \link{events_v2}}

\item{cluster_fronts_threshold_v2}{arguments for function \link{events_v2}}

\item{no_front_v2}{no_front argument for function \link{events_v2}}
}
\description{
Fetches the raw data using a connection to the database based on dbcon, then calculates when an individual was assumably at the nest box. Runs across the databases SNBatWESTERHOLZ and SNBatWESTERHOLZ_v2. Uses parallel computing.
}
\note{
Note that currently you need to have saved your credentials via save_credentials in order to run the function.
}
\examples{
df = data.table(box = 1, from = "2016-05-01 00:00:00", to = "2019-06-30 00:00:00")
hooray = eva('lschlicht', df, time_threshold_v1 = 10, tr_threshold_v2 = 100)
hooray
plot(hooray)
plot(subset(hooray, !is.na(transp) & box == 1 & in_ > "2019-03-15" & in_ < "2019-04-05"))
}
\author{
LS
}
