% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_EVENTS_elements_v1.R
\name{fetch_ins_outs_v1}
\alias{fetch_ins_outs_v1}
\title{Fetch individual ins and outs}
\usage{
fetch_ins_outs_v1(x, time_threshold)
}
\arguments{
\item{x}{a data.frame as provided by the function trim_data}
}
\value{
a data.frame where individual ins and outs are grouped together
}
\description{
This function fetches individual in and out events based on a time threshold and transponder readings.
If two transponder readings occur close in time, the rows in between are removed, because it is unclear to which
individual this data belongs.
}
\examples{
see function events
}
\seealso{
\code{\link{events}}, \code{\link{fetch_data}}, \code{\link{remove_faulty_data}},
\code{\link{trim_data}}, \code{\link{assign_direction}}, \code{\link{concat_events}},
}
\author{
LS
}
