% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_butler_ELEMENTS.R
\name{file_copy_status}
\alias{file_copy_status}
\alias{card_copy_status}
\title{Checks inserted SD cards.}
\usage{
file_copy_status(download_date = Sys.Date(), x = cardReader(), ...)

card_copy_status(x = file_copy_status())
}
\arguments{
\item{download_date}{default to Sys.Date()}

\item{x}{a DT as returned by cardReader() for file_copy_status
or a DT returned by file_copy_status() for card_copy_status()}

\item{...}{pass to  read.boxnumber, specifically hwidCheck}
}
\value{
a DT
}
\description{
Checks inserted SD cards.
}
\examples{
\dontrun{
 require(SNB2)
 bid <- harwareIDs()
 x = file_copy_status(boxid = bid)
 card_copy_status(x)
}
}
\author{
MV
}
