% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_rawtxt_to_db_ELEMENTS.R
\name{file_status_update1}
\alias{file_status_update1}
\title{File_status table: initial update}
\usage{
file_status_update1(con, x, update = TRUE)
}
\arguments{
\item{con}{a connection object}

\item{x}{a table exported by the incoming_files}

\item{update}{default to TRUE}
}
\value{
TRUE or the update content when update = FALSE
}
\description{
box, id, path,  filesize are saved to file_status table.
                 The id is a primary key in file_status and identifies chunks of data from the same file in each box b00x table.
}
\examples{
require(sdb)
con = dbcon('valcu')
x = incoming_files(con )
o = file_status_update1(update = FALSE)
dbDisconnect(con)
}
\author{
MV
}
