% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_installed}
\alias{find_installed}
\title{Find if an external software is installed}
\usage{
find_installed(nam)
}
\arguments{
\item{nam}{package name}
}
\value{
logical vector
}
\examples{
\dontrun{
find_installed('secure delete')
}
}
\author{
MV
}
