% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ELEMENTS_ds_to_scidb.R
\name{hot_files}
\alias{hot_files}
\title{Hot txt files (v1, v2)}
\usage{
hot_files(con, p = getOption("path.to.raw"), db = getOption("snbDB"))
}
\arguments{
\item{con}{A connection object}
}
\value{
\code{data.table}
}
\description{
Files listed in \code{file_status} table (upload_status is NULL) which are ready to process and upload
}
\author{
MV
}
