% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_rawtxt_to_db_ELEMENTS.R
\name{incoming_files}
\alias{incoming_files}
\title{Find out new data files}
\usage{
incoming_files(con, p = getOption("path.to.raw_v2"), y = year(Sys.Date()),
  db = getOption("snbDB_v2"))
}
\arguments{
\item{con}{a connection object}

\item{p}{path to raw data, default to getOption("path.to.raw")}

\item{y}{year, default to current year}
}
\value{
a data.table if there are new directories or otherwise NULL
}
\description{
The function lists new files which are not yet in  \code{SNBatWESTERHOLZ)v2.file_status}.
}
\section{Warning}{
 Before performing the file listing the function only looks for new directories therefore any
                   files which are copied to a given directory AFTER a db update will be overlooked!
}

\examples{
require(sdb)
con = dbcon('valcu')
incoming_files(con)
dbDisconnect(con)
}
\author{
MV
}
