% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_queries.R
\name{overnight}
\alias{overnight}
\title{overnight}
\usage{
overnight(buffer = 1, date = Sys.Date() - 1, ...)
}
\arguments{
\item{buffer}{(hours)   sunrise + buffer}

\item{date}{date, default to last date in file_status table}

\item{...}{goes to dbqSNB}

\item{username}{username}

\item{host}{host}
}
\description{
overnight
}
\examples{
overnight(buffer = 2)
x = overnight( anydate('2018.03.06') )

}
