% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_butler_ELEMENTS.R
\name{read.boxnumber}
\alias{read.boxnumber}
\title{looks up the box number}
\usage{
read.boxnumber(path, boxid, hwidCheck = TRUE)
}
\arguments{
\item{path}{file path}

\item{boxid}{data.table with current box - hwid}

\item{hwidCheck}{check hwid}
}
\value{
data.table containing Boxnummer.txt info
}
\description{
looks up the box number in BoxNNNN.txt and optionally cross-checks it with the HW-ID
}
\examples{
\dontrun{
 require(SNB2)
 boxid <- harwareIDs()
 x = cardReader()
 z = x[, read.boxnumber(mountpoint, boxid = boxid), by = .(row.names(x)) ]
}

}
\author{
MV
}
