% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fileop.R
\name{read_snb_txt_v2}
\alias{read_snb_txt_v2}
\title{read SNB data}
\usage{
read_snb_txt_v2(f)
}
\arguments{
\item{f}{path to the snb file}
}
\value{
a  \code{data.table} .
}
\description{
read an SNB file as data.table-s (only transponder and Light barrier data are parsed).
}
\examples{
\dontrun{
x = read_snb_txt_v2(f = "/ds/raw_data_kemp/FIELD/Westerholz/SNB/RAWDATA_v2/2016/2016.06.07/267/BOX0267.TXT")
x = read_snb_txt_v2(f = "/ds/raw_data_kemp/FIELD/Westerholz/SNB/RAWDATA_v2/2017/2017.04.24/30/BOX0030.TXT")
}
}
\author{
MV
}
