% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_rawtxt_to_db_PIPELINE.R
\name{scidb_snbUpdater}
\alias{scidb_snbUpdater}
\title{Box files (txt) to scidb.mpio.orn.mpg.de}
\usage{
scidb_snbUpdater(con, p = getOption("path.to.raw_v2"), y = year(Sys.Date()),
  db = getOption("snbDB_v2"), ui = FALSE, demo = FALSE, ...)
}
\arguments{
\item{con}{a connection object to \code{scidb.mpio.orn.mpg.de} the MySQL database.}

\item{ui}{only relevant in an UI interface (e.g. butler() )}

\item{demo}{only relevant when using a demo system (see install_demo_system() )}

\item{...}{goes to load_clean_txt: mainly parallel = TRUE}
}
\description{
scidb_snbUpdater serializes several functions and performs the update of both the file_status and each b00x table.
}
\examples{
\\donotrun{
 install_demo_system('mihai', '127.0.0.1')
 require(SNB2)
 con = dbcon(user =  getOption("DB_user") , host = getOption("host"))
 scidb_snbUpdater(con,  parallel = TRUE)
 closeCon(con)
}
}
\author{
MV
}
