% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_TXT_to_b000.R
\name{scidb_snbUpdater}
\alias{scidb_snbUpdater}
\title{txt Box files (txt) to db b000 tables pipeline}
\usage{
scidb_snbUpdater()
}
\arguments{
\item{u}{user, default to getOption("DB_user")}

\item{h}{host, default to getOption("host")}

\item{p}{path, default to getOption("path.to.raw_v2")}

\item{y}{year, default to year(Sys.Date()}

\item{db}{db name, default to getOption("snbDB_v2")}
}
\value{
TRUE if all txt are updated
}
\examples{
\dontrun{
 require(SNB2)
 demo_setup()
 scidb_snbUpdater()
}

}
\author{
MV
}
