% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_rawtxt_to_db_PIPELINE.R
\name{scidb_snbUpdater}
\alias{scidb_snbUpdater}
\title{Box files (txt) to scidb.mpio.orn.mpg.de}
\usage{
scidb_snbUpdater(u = getOption("DB_user"), h = getOption("host"),
  p = getOption("path.to.raw_v2"), y = year(Sys.Date()),
  db = getOption("snbDB_v2"))
}
\arguments{
\item{u}{user, default to getOption("DB_user")}

\item{h}{host, default to getOption("host")}

\item{p}{path, default to getOption("path.to.raw_v2")}

\item{y}{year, default to year(Sys.Date()}

\item{db}{db name, default to getOption("snbDB_v2")}
}
\value{
TRUE if update_bTables() is called and successful
}
\description{
scidb_snbUpdater serializes several functions and performs the 
               update of both the file_status and each b00x table.
}
\examples{
\\donotrun{
 require(SNB2)
 install_demo_SNB('mihai', '127.0.0.1')
 scidb_snbUpdater()
}

}
\author{
MV
}
