% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_unattended_pipelines.R
\name{scidb_snbUpdater}
\alias{scidb_snbUpdater}
\title{Unattended pipelines}
\usage{
scidb_snbUpdater(file = "~/scidb_snbUpdater.log")
}
\examples{
\dontrun{
 require(SNB2)
 
 demo_setup(install.test.db = TRUE, admin.user = 'mihai')
 scidb_snbUpdater()
}


}
\author{
MV
}
