% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_butler_ELEMENTS.R
\name{sdcard_format}
\alias{sdcard_format}
\title{Format SD cards}
\usage{
sdcard_format(password, x = card_format_status())
}
\arguments{
\item{password}{password}

\item{x}{card_format_status() output}
}
\author{
MV
}
