% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_butler_ELEMENTS.R
\name{sdcard_uploader}
\alias{sdcard_uploader}
\title{Copy SD cards content to repository}
\usage{
sdcard_uploader(x)
}
\arguments{
\item{x}{file_copy_status output}
}
\value{
TRUE on success
}
\examples{
 require(sdb)
 boxid <- dbq( q= 'select box, hwid from boxid order by box, datetime_ desc',
         db = getOption('snbDB_v2'), host = getOption('host'), user =getOption('DB_user'))[!duplicated(box)]
 x = file_copy_status(bid = boxid)
 sdcard_uploader(x)
}
\author{
MV
}
