% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_butler_ELEMENTS.R
\name{sdcard_uploader}
\alias{sdcard_uploader}
\title{Copy SD cards content to repository}
\usage{
sdcard_uploader(x)
}
\arguments{
\item{x}{file_copy_status output}
}
\value{
TRUE on success
}
\description{
Copy SD cards content to repository
}
\examples{
\dontrun{
 require(SNB2)
 boxid <- harwareIDs()
 x = file_copy_status(bid = boxid)
 sdcard_uploader(x)
}
}
\author{
MV
}
