% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_v2.R
\name{snbstring2date_v2}
\alias{snbstring2date_v2}
\title{Extract date}
\usage{
snbstring2date_v2(x)
}
\arguments{
\item{x}{character vector(s)}
}
\value{
\code{POSIXct} object
}
\description{
Extract date from raw data string
}
\examples{
  char2date(x = '20160526-123446 Transponder: 4B76C4B43A6F0001')
}
\author{
MV
}
