% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_rawtxt_to_db_ELEMENTS.R
\name{update_bTables}
\alias{update_bTables}
\title{Update  box and file_status tables}
\usage{
update_bTables(con, dat, db = getOption("snbDB_v2"))
}
\arguments{
\item{con}{a connection object}

\item{dat}{a list of \code{data.table}-s with attributes (see \code{load_clean_txt_v2} ).}
}
\value{
TRUE when both boxes (all) and  file status was updated else FALSE.
}
\description{
Update  box tables with parsed txt data (e.g. by \code{load_clean_txt_v2}  )
}
\author{
MV
}
