% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_ELEMENTS_ds_to_scidb.R
\name{well_formated_directory}
\alias{well_formated_directory}
\title{check the raw data directory (v1, v2)}
\usage{
well_formated_directory(p = getOption("path.to.raw"), y = year(Sys.Date()))
}
\arguments{
\item{p}{path to raw data, default to getOption("path.to.raw")}
}
\value{
logical
}
\description{
The function checks the proper directory format: \code{ year/date[prefix]/box/b00x.txt}
}
\examples{
# v1
well_formated_directory()
# v2
well_formated_directory(getOption("path.to.raw_v2"))
}
\author{
MV
}
