% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fileop.R
\name{well_formated_directory}
\alias{well_formated_directory}
\title{Check the raw data directory}
\usage{
well_formated_directory(p = getOption("path.to.raw_v2"),
  y = year(Sys.Date()))
}
\arguments{
\item{p}{path to raw data, default to getOption("path.to.raw")}
}
\value{
TRUE on success; the offending paths on failure.
}
\description{
The function checks the proper directory format: 
                    \code{ year/date[prefix]/box/b000.txt}
}
\examples{
\dontrun{
well_formated_directory(getOption("path.to.raw_v2"))
}
}
\author{
MV
}
